Refactor selection to always use system commands
