+++ crates/viewer/re_context_menu/src/actions/add_entities_to_new_view.rs
@@ -5,7 +5,7 @@ use nohash_hasher::IntSet;
 use re_log_types::{EntityPath, EntityPathFilter, EntityPathRule, RuleEffect};
 use re_types::ViewClassIdentifier;
 use re_ui::UiExt as _;
-use re_viewer_context::{Item, RecommendedView};
+use re_viewer_context::{Item, RecommendedView, SystemCommand, SystemCommandSender as _};
 use re_viewport_blueprint::ViewBlueprint;
 
 use crate::{ContextMenuAction, ContextMenuContext};
@@ -174,8 +174,8 @@ fn create_view_for_selected_entities(
     ctx.viewport_blueprint
         .add_views(std::iter::once(view), target_container_id, None);
     ctx.viewer_context
-        .selection_state()
-        .set_selection(Item::View(view_id));
+        .command_sender()
+        .send_system(SystemCommand::SetSelection(Item::View(view_id).into()));
     ctx.viewport_blueprint
         .mark_user_interaction(ctx.viewer_context);
 }
