+++ crates/viewer/re_global_context/src/lib.rs
@@ -8,6 +8,7 @@ mod command_sender;
 mod contents;
 mod file_dialog;
 mod item;
+mod item_collection;
 mod recording_or_table;
 
 pub use self::{
@@ -19,6 +20,7 @@ pub use self::{
     contents::{Contents, ContentsName, blueprint_id_to_tile_id},
     file_dialog::santitize_file_name,
     item::{Item, resolve_mono_instance_path, resolve_mono_instance_path_item},
+    item_collection::{ItemCollection, ItemContext},
     recording_or_table::RecordingOrTable,
 };
